Add docs for Icon usage and include in storybook
